Change fullview to view and view to quiz, and ensures answer is one of the choices
