+++ src/main/java/seedu/flashcard/logic/LogicManager.java
@@ -9,7 +9,7 @@ import seedu.flashcard.commons.core.GuiSettings;
 import seedu.flashcard.commons.core.LogsCenter;
 import seedu.flashcard.logic.commands.Command;
 import seedu.flashcard.logic.commands.CommandResult;
-import seedu.flashcard.logic.commands.ViewCommand;
+import seedu.flashcard.logic.commands.QuizCommand;
 import seedu.flashcard.logic.commands.exceptions.CommandException;
 import seedu.flashcard.logic.parser.FlashcardListParser;
 import seedu.flashcard.logic.parser.exceptions.ParseException;
@@ -40,7 +40,7 @@ public class LogicManager implements Logic {
         logger.info("----------------[USER COMMAND][" + commandText + "]");
         Command command = flashcardListParser.parseCommand(commandText);
         CommandResult commandResult = command.execute(model);
-        if (!(command instanceof ViewCommand)) {
+        if (!(command instanceof QuizCommand)) {
             model.updateLastViewedFlashcard(null);
         }
         try {
