+++ src/main/java/seedu/flashcard/logic/commands/QuizCommand.java
@@ -10,32 +10,32 @@ import seedu.flashcard.model.Model;
 import seedu.flashcard.model.flashcard.Flashcard;
 
 /**
- * Command to view the MCQ choicecs as well as the answer in a flashcard.
+ * Command to view the MCQ choices in a flashcard. The answer will not be shown.
  */
-public class FullViewCommand extends Command {
+public class QuizCommand extends Command {
 
-    public static final String COMMAND_WORD = "fullview";
+    public static final String COMMAND_WORD = "quiz";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
         + ": View the flashcard identified by the index number used in the displayed flashcard list.\n"
-        + "Both the choices and the answer will be shown.\n"
+        + "Only the choices with the flashcard will be shown but the answer will not.\n"
         + "Parameters: INDEX (must be a positive integer)\n"
         + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_INVALID_FLASHCARD_INDEX = "The index you entered is invalid!";
-    public static final String MESSAGE_DELETE_FLASHCARD_SUCCESS = "The flashcard is shown in full below.";
+    public static final String MESSAGE_DELETE_FLASHCARD_SUCCESS = "The flashcard is shown below.";
     private final Index targetIndex;
 
-    public FullViewCommand(Index targetIndex) {
+    public QuizCommand(Index targetIndex) {
         requireNonNull(targetIndex);
         this.targetIndex = targetIndex;
     }
 
     /**
-     * Showing a whole flashcard in the command bot with the answer.
-     * @param model {@code Model} which the command should operate on.
-     * @return the execution result containing the full flashcard.
-     * @throws CommandException error encountered during execution of the command.
+     * Showing a whole flashcard in the command bot, the answer will not be shown.
+     * @param model list of flashcards
+     * @return the execution result containing the flashcard without the answer.
+     * @throws CommandException error encountered during execution of command.
      */
     @Override
     public CommandResult execute(Model model) throws CommandException {
@@ -46,13 +46,13 @@ public class FullViewCommand extends Command {
         }
         Flashcard cardToView = lastShownList.get(targetIndex.getZeroBased());
         model.updateLastViewedFlashcard(cardToView);
-        return new CommandResult(cardToView.fullString());
+        return new CommandResult(cardToView.toString());
     }
 
     @Override
     public boolean equals(Object other) {
         return other == this
-            || (other instanceof FullViewCommand
-            && targetIndex.equals(((FullViewCommand) other).targetIndex));
+            || (other instanceof QuizCommand
+            && targetIndex.equals(((QuizCommand) other).targetIndex));
     }
 }
