+++ src/main/java/seedu/flashcard/logic/commands/ViewCommand.java
@@ -10,7 +10,7 @@ import seedu.flashcard.model.Model;
 import seedu.flashcard.model.flashcard.Flashcard;
 
 /**
- * Command to view the MCQ choices in a flashcard. The answer will not be shown.
+ * Command to view the MCQ choicecs as well as the answer in a flashcard.
  */
 public class ViewCommand extends Command {
 
@@ -18,12 +18,12 @@ public class ViewCommand extends Command {
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
         + ": View the flashcard identified by the index number used in the displayed flashcard list.\n"
-        + "Only the choices with the flashcard will be shown but the answer will not.\n"
+        + "Both the choices and the answer will be shown.\n"
         + "Parameters: INDEX (must be a positive integer)\n"
         + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_INVALID_FLASHCARD_INDEX = "The index you entered is invalid!";
-    public static final String MESSAGE_DELETE_FLASHCARD_SUCCESS = "The flashcard is shown below.";
+    public static final String MESSAGE_DELETE_FLASHCARD_SUCCESS = "The flashcard is shown in full below.";
     private final Index targetIndex;
 
     public ViewCommand(Index targetIndex) {
@@ -32,10 +32,10 @@ public class ViewCommand extends Command {
     }
 
     /**
-     * Showing a whole flashcard in the command bot, the answer will not be shown.
-     * @param model list of flashcards
-     * @return the execution result containing the flashcard without the answer.
-     * @throws CommandException error encountered during execution of command.
+     * Showing a whole flashcard in the command bot with the answer.
+     * @param model {@code Model} which the command should operate on.
+     * @return the execution result containing the full flashcard.
+     * @throws CommandException error encountered during execution of the command.
      */
     @Override
     public CommandResult execute(Model model) throws CommandException {
@@ -46,7 +46,7 @@ public class ViewCommand extends Command {
         }
         Flashcard cardToView = lastShownList.get(targetIndex.getZeroBased());
         model.updateLastViewedFlashcard(cardToView);
-        return new CommandResult(cardToView.toString());
+        return new CommandResult(cardToView.fullString());
     }
 
     @Override
