+++ src/main/java/seedu/flashcard/logic/parser/ViewCommandParser.java
@@ -3,13 +3,13 @@ package seedu.flashcard.logic.parser;
 import static seedu.flashcard.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.flashcard.commons.core.index.Index;
-import seedu.flashcard.logic.commands.ViewCommand;
+import seedu.flashcard.logic.commands.QuizCommand;
 import seedu.flashcard.logic.parser.exceptions.ParseException;
 
 /**
- * Parses input argument to generate a {@code ViewCommand}
+ * Parses input argument to generate a {@code QuizCommand}
  */
-public class ViewCommandParser implements Parser<ViewCommand> {
+public class ViewCommandParser implements Parser<QuizCommand> {
 
     /**
      * Parses the given context into a view command
@@ -17,12 +17,12 @@ public class ViewCommandParser implements Parser<ViewCommand> {
      * @return new view command
      */
     @Override
-    public ViewCommand parse(String args) throws ParseException {
+    public QuizCommand parse(String args) throws ParseException {
         try {
             Index index = ParserUtil.parseIndex(args);
-            return new ViewCommand(index);
+            return new QuizCommand(index);
         } catch (ParseException e) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, ViewCommand.MESSAGE_USAGE), e);
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, QuizCommand.MESSAGE_USAGE), e);
         }
     }
 }
